+++ packages/PsMenuKit/src/Core/PsMenuKit.Core.psm1
@@ -22,4 +22,5 @@ Export-ModuleMember -Function @(
     'New-PsMenu'
     'New-PsMenuItem'
     'Show-PsMenu'
+    'Get-PsMenuDisplayText'
 )